Migrate fjs/protocol/json_rpc to .f.mjs
Dependency-leaf implementation and proof migrate from authored
TypeScript to JSDoc-typed JavaScript under the stage-1 migration.
Id, Request, RpcError, Response, Handler, and Handlers move to a
sibling types.ts since fjs/protocol/mcp/module,
fjs/protocol/mcp/stdio/module, and their proofs depend on them.
Updates all four dependents' import specifiers and fixes stale
module.f.ts doc references (line-number citations in
fjs/protocol/json_rpc/todo/*.md, prose references in
fjs/protocol/mcp/todo/README.md).
