+++ fjs/mcp/proof.f.ts
@@ -7,7 +7,7 @@ import { create } from '../effects/memory/module.f.mjs'
 import { parse as parseJson } from '../media/json/module.f.mjs'
 import { number as rttiNumber, option, string as rttiString } from '../types/rtti/module.f.mjs'
 import { parse as rttiParse } from '../types/rtti/parse/module.f.mjs'
-import type { Response } from '../protocol/json_rpc/module.f.ts'
+import type { Response } from '../protocol/json_rpc/types.ts'
 import type { Vec } from '../types/bit_vec/types.ts'
 import { msb, u8ListToVec, vec8, repeat, length, maxLengthBytes } from '../types/bit_vec/module.f.mjs'
 import { vecToCBase32 } from '../basen/cbase32/module.f.mjs'
